requirements.txt
arrow==0.10.0
beautifulsoup4==4.5.3
binaryornot==0.4.0
bottle==0.12.17
bottle-sqlalchemy==0.4.3
chardet==2.3.0
click==6.7
cookiecutter==1.5.1
cookiejar==0.0.1
cycler==0.10.0
fake-useragent==0.1.4
future==0.16.0
gevent==1.2.1
greenlet==0.4.12
grequests==0.3.0
Jinja2==2.10.1
jinja2-time==0.2.0
lxml==3.7.2
MarkupSafe==0.23
matplotlib==2.0.0
numpy==1.12.0
pager==3.3
pandas==0.19.2
poyo==0.4.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
scikit-learn==0.18.1
scipy==0.18.1
scrapely==0.13.3
simplejson==3.10.0
six==1.10.0
sklearn==0.0
SQLAlchemy==1.3.8
tqdm==4.36.1
urllib3==1.20
w3lib==1.16.0
whichcraft==0.4.0
zxcvbn==4.4.28
